big resharper refactor
