+++ Doug/Controllers/CasinoController.cs
@@ -1,9 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Doug.Commands;
-using Microsoft.AspNetCore.Http;
+using Doug.Controllers.Dto;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Doug.Controllers
