+++ Doug/Models/DougContext.cs
@@ -1,9 +1,5 @@
 ﻿using Doug.Items;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace Doug.Models
 {
