+++ Doug/Models/Secret.cs
@@ -1,10 +1,4 @@
-﻿using Doug.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Doug.Models
+﻿namespace Doug.Models
 {
     public class Secret
     {
