+++ Doug/Models/Slur.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Doug.Models
+﻿namespace Doug.Models
 {
     public class Slur
     {
