+++ Doug/Models/User.cs
@@ -1,8 +1,4 @@
-﻿using Doug.Items;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 
 namespace Doug.Models
 {
