+++ Doug/Repositories/CoffeeRepository.cs
@@ -1,8 +1,6 @@
 ﻿using Doug.Models;
-using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
 
 namespace Doug.Repositories
 {
