+++ Doug/Services/AuthorizationService.cs
@@ -1,7 +1,4 @@
 ﻿using Doug.Slack;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 
 namespace Doug.Services
