+++ Doug/Services/EventService.cs
@@ -1,9 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Doug.Models;
-using Doug.Slack;
 
 namespace Doug.Services
 {
