+++ Doug/Slack/Dto/UserInfoResponse.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Doug.Slack
+﻿namespace Doug.Slack.Dto
 {
     public class UserInfoResponse
     {
