+++ Doug/Slack/SlackWebApi.cs
@@ -1,11 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
 using System.Web;
 using Doug.Models;
 using Doug.Repositories;
+using Doug.Slack.Dto;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
